+++ paper/paper.bib
@@ -1,16 +1,16 @@
 @article{chen2013knn,
-  author={Q. {Chen} and D. {Li} and C. {Tang}},
-  journal={IEEE Transactions on Pattern Analysis and Machine Intelligence}, 
-  title={KNN Matting}, 
-  year={2013},
-  volume={35},
-  number={9},
-  pages={2175-2188},
-  abstract={This paper proposes to apply the nonlocal principle to general alpha matting for the simultaneous extraction of multiple image layers; each layer may have disjoint as well as coherent segments typical of foreground mattes in natural image matting. The estimated alphas also satisfy the summation constraint. As in nonlocal matting, our approach does not assume the local color-line model and does not require sophisticated sampling or learning strategies. On the other hand, our matting method generalizes well to any color or feature space in any dimension, any number of alphas and layers at a pixel beyond two, and comes with an arguably simpler implementation, which we have made publicly available. Our matting technique, aptly called KNN matting, capitalizes on the nonlocal principle by using K nearest neighbors (KNN) in matching nonlocal neighborhoods, and contributes a simple and fast algorithm that produces competitive results with sparse user markups. KNN matting has a closed-form solution that can leverage the preconditioned conjugate gradient method to produce an efficient implementation. Experimental evaluation on benchmark datasets indicates that our matting results are comparable to or of higher quality than state-of-the-art methods requiring more involved implementation. In this paper, we take the nonlocal principle beyond alpha estimation and extract overlapping image layers using the same Laplacian framework. Given the alpha value, our closed form solution can be elegantly generalized to solve the multilayer extraction problem. We perform qualitative and quantitative comparisons to demonstrate the accuracy of the extracted image layers.},
-  keywords={conjugate gradient methods;feature extraction;image colour analysis;image matching;image segmentation;KNN matting;nonlocal principle;general alpha matting;multiple image layers;coherent segments;natural image matting;summation constraint;nonlocal matting;feature space;k nearest neighbors;nonlocal neighborhoods matching;sparse user markups;closed-form solution;preconditioned conjugate gradient method;benchmark datasets;Laplacian framework;multilayer extraction problem;qualitative comparisons;quantitative comparisons;Kernel;Laplace equations;Image color analysis;Materials;Vectors;Mathematical model;Image segmentation;Natural image matting;layer extraction},
-  doi={10.1109/TPAMI.2013.18},
-  ISSN={1939-3539},
-  month={Sep.}
+    author={Q. {Chen} and D. {Li} and C. {Tang}},
+    journal={IEEE Transactions on Pattern Analysis and Machine Intelligence},
+    title={KNN Matting},
+    year={2013},
+    volume={35},
+    number={9},
+    pages={2175-2188},
[…]
