### 2020-07-04 f514396
Fix typo in comment in ichol.py
