### 2020-07-10 7a13620
Add multi-level foreground estimation bibtex entry
