+++ doc/source/pymatting.bib
@@ -213,12 +213,8 @@
 }
 
 @article{germer2020multilevel,
-	title={Fast Multilevel Foreground Extraction},
-	author={Germer, Thomas and Uelwer, Tobias and Conrad, Stefan and Harmeling, Stefan},
-	journal={To be published},
-	volume={},
-	number={},
-	pages={},
-	year={2020},
-	publisher={}
+  title={Fast Multi-Level Foreground Estimation},
+  author={Germer, Thomas and Uelwer, Tobias and Conrad, Stefan and Harmeling, Stefan},
+  journal={arXiv preprint arXiv:2006.14970},
+  year={2020}
 }
