### 2020-07-13 #2423 @gkthiruvathukal
@macrocosme Would you be willing to contribute a review for this JOSS submission? I need a second, since this is outside @faroit's expertise.

@faroit Even when not entirely your expertise, the goal is to know whether the software meets the JOSS criteria and can perform its mission successfully. I hope you'd be willing to be the second reviewer, if @macrocosme is not available.
