### 2020-07-13 #2481 @whedon
[REVIEW]: PyMatting: A Python Library for Alpha Matting
**Submitting author:** @99991 (<a href="http://orcid.org/0000-0003-1355-7724">Thomas Germer</a>)
**Repository:** <a href="https://github.com/pymatting/pymatting" target ="_blank">https://github.com/pymatting/pymatting</a>
**Version:** v1.0.7
