### 2020-07-16 1fda617
Update CONTRIBUTING.md
