### 2020-07-16 614c263
Move community guidelines to CONTRIBUTING.md
