+++ CONTRIBUTING.md
@@ -1,15 +1,17 @@
-# Bug Reports
+# Community Guidelines
+
+## Bug Reports
 For bug reports please open a new issue. Include a [minimal reproducible example](https://stackoverflow.com/help/minimal-reproducible-example) and state your operating system, as well as the versions of all relevant Python packages you are using.
 
 **Important:** Before opening a new issue, make sure to check wether a similar issue already exists.
 
-# Questions
+## Questions
 
 If you have any other questions about the usage of the library feel free to open an issue or chat with us on [gitter](https://gitter.im/pymatting/community).
 
 **Important:** Before opening a new issue, make sure to check wether a similar issue already exists.
 
-# Contributing Source Code
+## Contributing Source Code
 
 Thank you for considering to contribute to the PyMatting project!
 
@@ -28,7 +30,7 @@ To make a contribution, follow those steps:
 
 For larger changes, it is probably a good idea to open an issue and discuss them first.
 
-## Licensing
+### Licensing
 
 * All contributed code must be released under the MIT license.
 * You must be legally allowed to release the contributed code under the MIT license.
