+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,38 @@
+---
+name: Bug report
+about: Create a report to help us improve the PyMatting library.
+title: "[BUG \U0001F41B]"
+labels: ''
+assignees: ''
+
+---
+
+**Bug description**
+
+(Bug description here)
+
+**To Reproduce**
+
+(Include a [minimal reproducible example](https://stackoverflow.com/help/minimal-reproducible-example), input/output images and instructions on how to run it.)
+
+**Expected behavior**
+
+(What should happen instead?)
+
+**Images**
+
+(Add relevant images.)
+
+**Library versions:**
+
+(Run the following commands and paste the result here.)
+
+```bash
+python --version --version
+python -c "import numpy; numpy.show_config()"
+python -c "import scipy;scipy.show_config()"
+python -c "import numba;print('Numba version:', numba.__version__)"
+python -c "import PIL;print('PIL version:', PIL.__version__)"
+python -c "from pymatting.__about__ import __version__;print('PyMatting version:', __version__)"
+
+```
