+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,10 @@
+---
+name: Question
+about: Ask us a general question about the PyMatting library.
+title: "[Question❓]"
+labels: ''
+assignees: ''
+
+---
+
+
