### 2020-07-16 #2481 @macrocosme
@99991 @gkthiruvathukal First of all, I would like to congratulate the authors on their clean and useful software. The code is well documented, includes tests, and benchmarks. The code was easily installable, and examples ran as expected (except for a few path issues that have now been fixed). Instructions on how to set and run the benchmarks have also been simple to follow and worked as expected. I have a few minor comments regarding the submission.

* Many Figures do not include axes labels. Labels are required for all Figures, even when described in the Figure caption.
* There is no hardware and OS description available describing how the benchmarks' results were obtained, which may be suitable.

Regarding  the [community guidelines criteria](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation):
* there are clear guidelines for third parties wishing to **contribute to the software**. I think it would be useful to add a mention in the README directing to the CONTRIBUTING.md file;
* Unless I missed it, there are no clear guidelines about how to **report issues or problems with the software**;
* Unless I missed it, there are no clear guidelines about how to **seek support**. However, I must note that there is the "chat on gitter" badge where it is clearly possible to do so.

I would like to ask the authors to apply these changes, after which I will accept this submission.
