### 2020-07-16 c61f6a2
Add hardware information
