+++ paper/paper.md
@@ -119,7 +119,7 @@ Our results are consistent with the results achieved by the authors' implementat
 
 ![Comparison of runtime for different image sizes.\label{fig:runtimes}](figures/time_image_size-crop.pdf)
 
-We compare the computational runtime of our solver with other solvers: PyAMG [@pyamg], UMFPACK [@umfpack], AMGCL [@amgcl], MUMPS [@MUMPS-a; @MUMPS-b], Eigen [@eigen] and SuperLU [@li1999superlu]. \autoref{fig:memory-runtime} and \autoref{fig:runtimes} show that our implemented conjugate gradient method in combination with the incomplete Cholesky decomposition preconditioner outperforms the other methods in terms of computational runtime by a large margin. For the iterative solver, we use an absolute tolerance of $10^{-7}$, which we scale with the number of known pixels, i.e., pixels that are either marked as foreground or background in the trimap. The benchmarked linear system arises from the matting Laplacian by @levin2007closed. \autoref{fig:memory-runtime} shows that our solver also outperforms the other solvers in terms of memory usage.
+We compare the computational runtime of our solver with other solvers: PyAMG [@pyamg], UMFPACK [@umfpack], AMGCL [@amgcl], MUMPS [@MUMPS-a; @MUMPS-b], Eigen [@eigen] and SuperLU [@li1999superlu]. \autoref{fig:memory-runtime} and \autoref{fig:runtimes} show that our implemented conjugate gradient method in combination with the incomplete Cholesky decomposition preconditioner outperforms the other methods in terms of computational runtime by a large margin. For the iterative solver, we use an absolute tolerance of $10^{-7}$, which we scale with the number of known pixels, i.e., pixels that are either marked as foreground or background in the trimap. The benchmarked linear system arises from the matting Laplacian by @levin2007closed. \autoref{fig:memory-runtime} shows that our solver also outperforms the other solvers in terms of memory usage. Benchmarks were performed on a high-performance computer with an Intel Xeon Gold 6134 CPU (3.20 GHz) and 196 GB memory running Ubuntu 16.04.
 
 
 ## Compatibility and Extendability
