### 2020-07-16 aa0012c
Update paper.md
