### 2020-07-16 9fb6acc
Update paper.md
