### 2020-07-16 #2481 @tuelwer
Thanks for the fast review and the helpful comments, @macrocosme! We really appreciate this.

Regarding your minor comments:

- We added axes labels for Figure 2 and Figure 3 in commit https://github.com/pymatting/pymatting/commit/845e5e82094ef99e65777abc0040969b3271ab36.
- Furthermore, we added hardware and OS description at the end of the performance comparison section in commit https://github.com/pymatting/pymatting/commit/c61f6a2b81e42478d393b16018539fa2ecff41c3.
- We updated the content of the [CONTRIBUTING.md](https://github.com/pymatting/pymatting/blob/master/CONTRIBUTING.md) file. Now we have guidelines for reporting bugs and requesting support. We also added issue templates to our repository.
