### 2020-08-20 #2481 @ziatdinovmax
My apologies for the delay in responding... Congrats to the authors on their very nice library! It installs and runs smoothly, and the documentation is in good order. The only minor comment that I have is about the paper. When I was reading it, I kept wondering (perhaps out of my ignorance) who is the target audience and what specific applications it can be used for. The statement at the beginning of the paper that it is “important for many image editing tasks” seems to be a bit vague.  Perhaps listing several potential applications e.g. in the field of scientific and/or medical imaging together with references will be helpful.
