+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@article{Fan2019hierarchical,
+@article{fan2019hierarchical,
     author={Z. {Fan} and J. {Lu} and C. {Wei} and H. {Huang} and X. {Cai} and X. {Chen}},
     journal={IEEE Transactions on Image Processing}, 
     title={A Hierarchical Image Matting Model for Blood Vessel Segmentation in Fundus Images}, 
@@ -13,7 +13,7 @@
     month={May}
 }
 
-@inproceedings{Zhaozheng2010understanding,
+@inproceedings{zhaozheng2010understanding,
     author={Yin, Zhaozheng and Li, Kang and Kanade, Takeo and Chen, Mei},
     editor={Jiang, Tianzi and Navab, Nassir and Pluim, Josien P. W. and Viergever, Max A.},
     title={Understanding the Optics to Aid Microscopy Image Segmentation},
@@ -28,7 +28,7 @@
 }
 
 
-@inproceedings{Kanade2011cell,
+@inproceedings{kanade2011cell,
     author={T. {Kanade} and Z. {Yin} and R. {Bise} and S. {Huh} and S. {Eom} and M. F. {Sandbothe} and M. {Chen}},
     booktitle={2011 IEEE Workshop on Applications of Computer Vision (WACV)}, 
     title={Cell image analysis: Algorithms, system and applications}, 
