+++ paper/paper.md
@@ -24,10 +24,11 @@ bibliography: paper.bib
 
 # Summary
 
-An important step of many image editing tasks is to extract specific objects
+A fundamental problem of many image processing tasks is the extraction of specific objects
 from an image in order to place them in a scene of a movie or compose them onto
 another background. Alpha matting describes the problem of separating the objects
 in the foreground from the background of an image given only a rough sketch.
+Besides for everyday image editing, alpha matting has been applied in medical image analysis [@fan2019hierarchical; @zhaozheng2010understanding; @kanade2011cell].
 We introduce the PyMatting package for the Python ecosystem which implements
 various approaches to solve the alpha matting problem. Our library is also
 able to extract the foreground of an image given the alpha matte.
