### 2020-08-24 8b5bd1b
Update paper.md
