+++ paper/paper.md
@@ -28,12 +28,13 @@ A fundamental problem of many image processing tasks is the extraction of specif
 from an image in order to place them in a scene of a movie or compose them onto
 another background. Alpha matting describes the problem of separating the objects
 in the foreground from the background of an image given only a rough sketch.
-Besides for everyday image editing, alpha matting has been applied in medical image analysis [@fan2019hierarchical; @zhaozheng2010understanding; @kanade2011cell].
+Besides everyday image editing, alpha matting has been applied to, e.g., medical image analysis 
+[@fan2019hierarchical; @zhaozheng2010understanding;  and microscopy image restoration [@kanade2011cell].
 We introduce the PyMatting package for the Python ecosystem which implements
 various approaches to solve the alpha matting problem. Our library is also
 able to extract the foreground of an image given the alpha matte.
+The target audience of our toolbox are researchers of image processing and computer vision.
 The implementation aims to be computationally efficient and easy to use.
-The target audience of our toolbox are researchers of image processing and computer vision. 
 
 For an image $I$ with foreground pixels $F$ and background pixels $B$,
 alpha matting asks to determine opacities $\alpha$, such that the equality
