### 2020-08-24 4cc9d5e
Update paper.md
