+++ paper/paper.md
@@ -29,7 +29,7 @@ from an image in order to place them in a scene of a movie or compose them onto
 another background. Alpha matting describes the problem of separating the objects
 in the foreground from the background of an image given only a rough sketch.
 Besides everyday image editing, alpha matting has been applied to, e.g., medical image analysis 
-[@fan2019hierarchical; @zhaozheng2010understanding;  and microscopy image restoration [@kanade2011cell].
+[@fan2019hierarchical; @zhaozheng2010understanding] and microscopy image restoration [@kanade2011cell].
 We introduce the PyMatting package for the Python ecosystem which implements
 various approaches to solve the alpha matting problem. Our library is also
 able to extract the foreground of an image given the alpha matte.
