### 2020-08-24 bdbc884
Update paper.md
