+++ paper/paper.md
@@ -46,7 +46,7 @@ using a foreground estimation method (\autoref{fig:grid}).
 
 We introduce the PyMatting toolbox for Python which implements various approaches to solve 
 the alpha matting problem. Our library is also able to extract the foreground of an image given the alpha matte.
-The target audience of our library are researchers of image processing and computer vision.
+The target audience are researchers of image processing and computer vision.
 The implementation aims to be computationally efficient and easy to use.
 
 
