### 2020-08-24 934ea3c
Update paper.md
