+++ paper/paper.bib
@@ -17,7 +17,7 @@
     author={Yin, Zhaozheng and Li, Kang and Kanade, Takeo and Chen, Mei},
     editor={Jiang, Tianzi and Navab, Nassir and Pluim, Josien P. W. and Viergever, Max A.},
     title={Understanding the Optics to Aid Microscopy Image Segmentation},
-    booktitle={Medical Image Computing and Computer-Assisted Intervention -- MICCAI 2010},
+    booktitle={Medical Image Computing and Computer-Assisted Intervention -- {M}{I}{C}{C}{A}{I} 2010},
     year={2010},
     publisher={Springer Berlin Heidelberg},
     address={Berlin, Heidelberg},
@@ -30,7 +30,7 @@
 
 @inproceedings{kanade2011cell,
     author={T. {Kanade} and Z. {Yin} and R. {Bise} and S. {Huh} and S. {Eom} and M. F. {Sandbothe} and M. {Chen}},
-    booktitle={2011 IEEE Workshop on Applications of Computer Vision (WACV)}, 
+    booktitle={2011 {I}{E}{E}{E} Workshop on Applications of Computer Vision ({W}{A}{C}{V})}, 
     title={Cell image analysis: Algorithms, system and applications}, 
     year={2011},
     volume={},
