+++ paper/paper.bib
@@ -105,7 +105,7 @@
 
 @inproceedings{zheng2009learning,
     author={Y. {Zheng} and C. {Kambhamettu}},
-    booktitle={2009 IEEE 12th International Conference on Computer Vision},
+    booktitle={2009 {I}{E}{E}{E} 12th International Conference on Computer Vision},
     title={Learning based digital matting},
     year={2009},
     volume={},
