+++ paper/paper.bib
@@ -17,7 +17,7 @@
     author={Yin, Zhaozheng and Li, Kang and Kanade, Takeo and Chen, Mei},
     editor={Jiang, Tianzi and Navab, Nassir and Pluim, Josien P. W. and Viergever, Max A.},
     title={Understanding the Optics to Aid Microscopy Image Segmentation},
-    booktitle={Medical Image Computing and Computer-Assisted Intervention -- {M}{I}{C}{C}{A}{I} 2010},
+    booktitle={Medical {I}mage {C}omputing and {C}omputer-{A}ssisted {I}ntervention -- {M}{I}{C}{C}{A}{I} 2010},
     year={2010},
     publisher={Springer Berlin Heidelberg},
     address={Berlin, Heidelberg},
@@ -30,7 +30,7 @@
 
 @inproceedings{kanade2011cell,
     author={T. {Kanade} and Z. {Yin} and R. {Bise} and S. {Huh} and S. {Eom} and M. F. {Sandbothe} and M. {Chen}},
-    booktitle={2011 {I}{E}{E}{E} Workshop on Applications of Computer Vision ({W}{A}{C}{V})}, 
+    booktitle={2011 {I}{E}{E}{E} {W}orkshop on {A}pplications of {C}omputer {V}ision ({W}{A}{C}{V})}, 
     title={Cell image analysis: Algorithms, system and applications}, 
     year={2011},
     volume={},
@@ -105,7 +105,7 @@
 
 @inproceedings{zheng2009learning,
     author={Y. {Zheng} and C. {Kambhamettu}},
-    booktitle={2009 {I}{E}{E}{E} 12th International Conference on Computer Vision},
+    booktitle={2009 {I}{E}{E}{E} 12th {I}nternational {C}onference on {C}omputer {V}ision},
     title={Learning based digital matting},
     year={2009},
     volume={},
@@ -121,7 +121,7 @@
 @inproceedings{grady2005random,
     title={Random walks for interactive alpha-matting},
     author={Grady, Leo and Schiwietz, Thomas and Aharon, Shmuel and Westermann, R{\"u}diger},
-    booktitle={Proceedings of VIIP},
+    booktitle={Proceedings of {V}{I}{I}{P}},
     volume={2005},
     pages={423--429},
     year={2005}
@@ -147,7 +147,7 @@
     issn = {0098-3500},
     url = {https://doi.org/10.1145/992200.992206},
     doi = {10.1145/992200.992206},
-    journal = {ACM Trans. Math. Softw.},
+    journal = {ACM Transactions on Mathematical Software},
     month = jun,
     pages = {196–199},
     numpages = {4},
@@ -210,7 +210,7 @@
 
 @inproceedings{he2010fast,
     author={K. {He} and J. {Sun} and X. {Tang}},
-    booktitle={2010 IEEE Computer Society Conference on Computer Vision and Pattern Recognition},
+    booktitle={2010 {I}{E}{E}{E} {C}omputer {S}ociety {C}onference on {C}omputer {V}ision and {P}attern {R}ecognition},
     title={Fast matting using large kernel matting Laplacian matrices},
     year={2010},
     volume={},
@@ -225,7 +225,7 @@
 
 @inproceedings{rhemann2009perceptually,
     author={C. {Rhemann} and C. {Rother} and J. {Wang} and M. {Gelautz} and P. {Kohli} and P. {Rott}},
-    booktitle={2009 IEEE Conference on Computer Vision and Pattern Recognition},
+    booktitle={2009 {I}{E}{E}{E} {C}onference on {C}omputer {V}ision and {P}attern {R}ecognition},
     title={A perceptually motivated online benchmark for image matting},
     year={2009},
     volume={},
@@ -249,7 +249,7 @@
[…]
