+++ paper/paper.bib
@@ -294,7 +294,7 @@
     issn = {0098-3500},
     url = {https://doi.org/10.1145/200979.200981},
     doi = {10.1145/200979.200981},
-    journal = {ACM Trans. Math. Softw.},
+    journal = {ACM Transactions on Mathematical Software},
     month = mar,
     pages = {5–17},
     numpages = {13},
