### 2020-08-24 dd2468b
Update paper.md
