+++ paper/paper.md
@@ -24,14 +24,12 @@ bibliography: paper.bib
 
 # Summary
 
-An important step of many image editing tasks is to extract specific objects
+A fundamental problem of many image processing tasks is the extraction of specific objects
 from an image in order to place them in a scene of a movie or compose them onto
 another background. Alpha matting describes the problem of separating the objects
 in the foreground from the background of an image given only a rough sketch.
-We introduce the PyMatting package for the Python ecosystem which implements
-various approaches to solve the alpha matting problem. Our library is also
-able to extract the foreground of an image given the alpha matte.
-The implementation aims to be computationally efficient and easy to use.
+Besides everyday image editing, alpha matting has been applied to medical image analysis 
+[@fan2019hierarchical; @zhaozheng2010understanding] and microscopy image restoration [@kanade2011cell].
 
 For an image $I$ with foreground pixels $F$ and background pixels $B$,
 alpha matting asks to determine opacities $\alpha$, such that the equality
@@ -45,6 +43,11 @@ rough classification of the input image into foreground, background and unknown
 pixels, to further constrain the problem. Subsequently, the foreground $F$ can be
 extracted from the input image $I$ and the previously computed alpha matte $\alpha$
 using a foreground estimation method (\autoref{fig:grid}).
+
+We introduce the PyMatting toolbox for Python which implements various approaches to solve 
+the alpha matting problem. Our library is also able to extract the foreground of an image given the alpha matte.
+The target audience are researchers of image processing and computer vision.
+The implementation aims to be computationally efficient and easy to use.
 
 
 ## Implemented Methods for Alpha Matting
