+++ paper/paper.bib
@@ -1,3 +1,49 @@
+@article{fan2019hierarchical,
+    author={Z. {Fan} and J. {Lu} and C. {Wei} and H. {Huang} and X. {Cai} and X. {Chen}},
+    journal={IEEE Transactions on Image Processing}, 
+    title={A Hierarchical Image Matting Model for Blood Vessel Segmentation in Fundus Images}, 
+    year={2019},
+    volume={28},
+    number={5},
+    pages={2367-2377},
+    abstract={In this paper, a hierarchical image matting model is proposed to extract blood vessels from fundus images. More specifically, a hierarchical strategy is integrated into the image matting model for blood vessel segmentation. Normally, the matting models require a user specified trimap, which separates the input image into three regions: the foreground, background, and unknown regions. However, creating a user specified trimap is laborious for vessel segmentation tasks. In this paper, we propose a method that first generates trimap automatically by utilizing region features of blood vessels, then applies a hierarchical image matting model to extract the vessel pixels from the unknown regions. The proposed method has low calculation time and outperforms many other state-of-art supervised and unsupervised methods. It achieves a vessel segmentation accuracy of 96.0%, 95.7%, and 95.1% in an average time of 10.72s, 15.74s, and 50.71s on images from three publicly available fundus image datasets DRIVE, STARE, and CHASE_DB1, respectively.},
+    keywords={blood vessels;eye;feature extraction;image segmentation;medical image processing;hierarchical image matting model;blood vessels extraction;trimap;input image separation;matting models;fundus images;blood vessel segmentation;time 50.71 s;time 10.72 s;time 15.74 s;Image segmentation;Feature extraction;Blood vessels;Biomedical imaging;Deformable models;Diseases;Retina;Image matting;hierarchical strategy;fundus;trimap;region features;segmentation;vessel},
+    doi={10.1109/TIP.2018.2885495},
+    ISSN={1941-0042},
+    month={May}
+}
+
+@inproceedings{zhaozheng2010understanding,
+    author={Yin, Zhaozheng and Li, Kang and Kanade, Takeo and Chen, Mei},
+    editor={Jiang, Tianzi and Navab, Nassir and Pluim, Josien P. W. and Viergever, Max A.},
+    title={Understanding the Optics to Aid Microscopy Image Segmentation},
+    booktitle={Medical {I}mage {C}omputing and {C}omputer-{A}ssisted {I}ntervention -- {M}{I}{C}{C}{A}{I} 2010},
+    year={2010},
+    publisher={Springer Berlin Heidelberg},
+    address={Berlin, Heidelberg},
+    pages={209--217},
+    abstract={Image segmentation is essential for many automated microscopy image analysis systems. Rather than treating microscopy images as general natural images and rushing into the image processing warehouse for solutions, we propose to study a microscope's optical properties to model its image formation process first using phase contrast microscopy as an exemplar. It turns out that the phase contrast imaging system can be relatively well explained by a linear imaging model. Using this model, we formulate a quadratic optimization function with sparseness and smoothness regularizations to restore the ``authentic'' phase contrast images that directly correspond to specimen's optical path length without phase contrast artifacts such as halo and shade-off. With artifacts removed, high quality segmentation can be achieved by simply thresholding the restored images. The imaging model and restoration method are quantitatively evaluated on two sequences with thousands of cells captured over several days.},
+    isbn={978-3-642-15705-9},
[…]
