### 2020-08-24 #2481 @tuelwer
@ziatdinovmax @gkthiruvathukal thank you for your fast responses! We added the references to the paper and included the statement about the target audience in the first section of the paper.

We also made some further changes to the first section and the formatting of the references.
