### 2020-09-14 d0d36e1
Some small changes in the formulation
