### 2020-10-12 #2481 @tuelwer
@gkthiruvathukal We created a [tagged release](https://github.com/pymatting/pymatting/releases/tag/v1.0.7) of our package. The version number is 1.0.7 and the Zenodo DOI is [`10.5281/zenodo.4081416`](https://zenodo.org/record/4081416).
