### 2020-10-13 #2481 @gkthiruvathukal
Thanks, @danielskatz. I think we're ready to go here!  I am still classically conditioned to my "early days" of JOSS where I wasn't doing the whedon accept.
